refactor: extract channel handling to own file
